Moved the Player namespace out of Core
Separated the Player-related scripts from the Core namespace to create a new assembly
