+++ Assets/Code/Scripts/Core/Core.asmdef
@@ -9,12 +9,13 @@
         "GUID:ba75b8ae5d38bcc40a65785047ee8055",
         "GUID:fbd1169f7467a544ea39d78b143851af",
         "GUID:aaf86c895f7b60142a81dd2723129b42",
+        "GUID:0b652c770aa3eb54fbc16c1b9dfee8bc",
         "GUID:b3e14a8079a870246b2ce11145d6e4d2",
-        "GUID:7a0578f70705022478366244226c17a4",
         "GUID:4307f53044263cf4b835bd812fc161a4",
         "GUID:7369a7f1b010248859b67afbd57b1122",
         "GUID:fe25561d224ed4743af4c60938a59d0b",
-        "GUID:5e0b455853cd340e0813387221dfd501"
+        "GUID:5e0b455853cd340e0813387221dfd501",
+        "GUID:7a0578f70705022478366244226c17a4"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
