+++ Assets/Code/Scripts/Core/Managers/PlayerManager.cs.meta
