+++ Assets/Code/Scripts/Player/Commands/FreezePositionCommand.cs
@@ -1,16 +1,14 @@
 //*******************************************************************************************
-// IdleCommand
+// FreezeCommand
 //*******************************************************************************************
-using KrillOrBeKrilled.Core.Commands.Interfaces;
-
-namespace KrillOrBeKrilled.Core.Commands {
+namespace KrillOrBeKrilled.Player.Commands {
     /// <summary>
-    /// Implements <see cref="ICommand"/> to execute the player action for standing idle.
+    /// Implements <see cref="ICommand"/> to freeze the player position.
     /// </summary>
-    public class IdleCommand : ICommand {
+    public class FreezeCommand : ICommand {
         /// Reference to the player Pawn to control.
         private readonly Pawn _controlledObject;
-        
+
         //========================================
         // Public Methods
         //========================================
@@ -18,14 +16,14 @@ namespace KrillOrBeKrilled.Core.Commands {
         #region Public Methods
 
         public void Execute() {
-            this._controlledObject.StandIdle();
+            this._controlledObject.FreezePosition();
         }
         
         /// <summary>
         /// Constructor to set references required for this command to act on a <see cref="Pawn"/>.
         /// </summary>
-        /// <param name="controlledObject"> The player <see cref="Pawn"/> associated with this idle command. </param>
-        public IdleCommand(Pawn controlledObject) {
+        /// <param name="controlledObject"> The player <see cref="Pawn"/> associated with this deploy command. </param>
+        public FreezeCommand(Pawn controlledObject) {
             this._controlledObject = controlledObject;
         }
