+++ Assets/Code/Scripts/Player/Commands/GlideCommand.cs
@@ -1,21 +1,19 @@
 //*******************************************************************************************
-// JumpCommand
+// GlideCommand
 //*******************************************************************************************
-using KrillOrBeKrilled.Core.Commands.Interfaces;
-
-namespace KrillOrBeKrilled.Core.Commands {
+namespace KrillOrBeKrilled.Player.Commands {
     /// <summary>
-    /// Implements <see cref="ICommand"/> to execute the player action for jumping.
+    /// Implements <see cref="ICommand"/> to execute the player action for gliding.
     /// </summary>
-    public class JumpCommand : ICommand {
+    public class GlideCommand : ICommand {
         /// Reference to the player Pawn to control.
         private readonly Pawn _controlledObject;
 
         /// The move input for this command.
         private readonly float _moveInput;
 
-        /// The jump force for this command.
-        private readonly float _jumpForce;
+        /// The speed multiplier to apply while gliding.
+        private readonly float _glideSpeedMultiplier;
 
         //========================================
         // Public Methods
@@ -28,17 +26,16 @@ namespace KrillOrBeKrilled.Core.Commands {
         /// </summary>
         /// <param name="controlledObject"> The player <see cref="Pawn"/> associated with this jump command. </param>
         /// <param name="moveInput"> The direction to move the player <see cref="Pawn"/> on command execution. </param>
-        /// <param name="jumpForce"> The jump force. </param>
+        /// <param name="glideSpeedMultiplier"> The multiplier for the glide speed. </param>
         /// ;
-        public JumpCommand(Pawn controlledObject, float moveInput, float jumpForce) {
+        public GlideCommand(Pawn controlledObject, float moveInput, float glideSpeedMultiplier) {
             this._controlledObject = controlledObject;
             this._moveInput = moveInput;
-            this._jumpForce = jumpForce;
+            this._glideSpeedMultiplier = glideSpeedMultiplier;
         }
 
         public void Execute() {
-            this._controlledObject.Move(this._moveInput);
-            this._controlledObject.Jump(this._jumpForce);
+            this._controlledObject.Glide(this._moveInput, this._glideSpeedMultiplier);
         }
 
         #endregion
