+++ Assets/Code/Scripts/Player/Commands/NullCommand.cs
@@ -1,9 +1,7 @@
 //*******************************************************************************************
 // NullCommand
 //*******************************************************************************************
-using KrillOrBeKrilled.Core.Commands.Interfaces;
-
-namespace KrillOrBeKrilled.Core.Commands {
+namespace KrillOrBeKrilled.Player.Commands {
     /// <summary>
     /// A Null Object Command that implements <see cref="ICommand"/> to do nothing.
     /// </summary>
