+++ Assets/Code/Scripts/Player/Player.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0b652c770aa3eb54fbc16c1b9dfee8bc
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
