+++ Assets/Code/Scripts/Player/TrapController.cs.meta
