+++ Assets/Code/Scripts/UI/ControlsUI.cs
@@ -1,4 +1,4 @@
-using KrillOrBeKrilled.Core.Player;
+using KrillOrBeKrilled.Player.Input;
 using UnityEngine;
 
 //*******************************************************************************************
