+++ Assets/Code/Scripts/UI/UI.asmdef
@@ -8,7 +8,8 @@
         "GUID:aaf86c895f7b60142a81dd2723129b42",
         "GUID:7a0578f70705022478366244226c17a4",
         "GUID:bd384bcb37274bc49ab45c2f4aed36de",
-        "GUID:34aa492b82754644eac2f903cd496268"
+        "GUID:34aa492b82754644eac2f903cd496268",
+        "GUID:0b652c770aa3eb54fbc16c1b9dfee8bc"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
